More Fixes for TCP
Made some more fixes in the TCP classes and the PrinterTabComposite in
order to remove erroneous responses

+++ src/org/craneprint/craneserver/tcp/TCPListenerThread.java
@@ -49,10 +49,10 @@ public class TCPListenerThread implements Runnable {
 					int type = getType(jo);
 					if(type == RequestType.JOB_COMPLETE){
 						//this.getQueueManager().toString();
-						this.getQueueManager().printComplete((int)(long)jo.get("printerId"));
+						this.getQueueManager().printComplete((int)jo.get("printerId"));
 					}
 					if(type == RequestType.GET_NEW_JOB){
-						this.getQueueManager().sendNextInQueue((int)(long)jo.get("printerId"));
+						this.getQueueManager().sendNextInQueue((int)jo.get("printerId"));
 					}
 					outToAgent.writeBytes("{\"resp\":\"success\"}\n");
 					//outToAgent.close();
